add OpenSubdiv

+++ meson_ln.py
@@ -6,8 +6,8 @@ HERE = pathlib.Path(__file__).parent
 
 def process(subprojects: pathlib.Path):
     print(subprojects)
-    src = subprojects / "packagefiles/usd"
-    dst = subprojects / "usd"
+    src = subprojects / "packagefiles"
+    dst = subprojects
 
     for f in src.glob("**/meson.build"):
         rel = f.relative_to(src)
@@ -15,9 +15,10 @@ def process(subprojects: pathlib.Path):
         if target.exists() or target.is_symlink():
             print(f"remove {target}")
             os.unlink(target)
-        print(f"{f}")
-        print(f"  => {target}")
-        os.symlink(f, target)
+        if target.parent.exists():
+            print(f"{f}")
+            print(f"  => {target}")
+            os.symlink(f, target)
 
 
 if __name__ == "__main__":
